+++ CHANGES.rst
@@ -9,6 +9,7 @@ Changelog
 - `split-records` filter now allows specifying the meta-data field in which to store the split name
 - the `tee` meta-filter can now forward or drop the incoming data based on a meta-data evaluation
 - added `sub-process` filter for processing data with sub-flow of filters, can be conditional based on meta-data evaluation
+- the `metadata-from-name` filter can work on the path now as well (must be present)
 
 
 0.0.13 (2025-07-15)